pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "form13f-aiagent"
version = "0.1.0"
description = "AI agent for querying SEC Form 13F institutional holdings data"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    # Core
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    # Note: asyncpg removed - using psycopg2-binary for better compatibility

    # HTTP & Async
    "httpx>=0.26.0",
    "tenacity>=8.2.3",

    # XML Parsing
    "lxml>=5.0.0",
    "beautifulsoup4>=4.12.0",

    # SQL Parsing & Validation
    "sqlparse>=0.4.4",

    # AI/LLM
    "litellm>=1.53.0",

    # Phase 8B - RAG System
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.1.0",

    # API
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "slowapi>=0.1.9",

    # CLI
    "typer>=0.9.0",
    "rich>=13.7.0",
    "tqdm>=4.66.0",

    # UI
    "streamlit>=1.30.0",
    "plotly>=5.18.0",

    # Authentication
    "supabase>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
ui = [
    # UI components - now in main dependencies for Railway deployment
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "ipython>=8.20.0",
    "jupyter>=1.0.0",
]

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = [
    "src/ingestion/edgar_client.py",  # Old stub - not used
    "src/ingestion/parser.py",  # Old stub - replaced by tsv_parser
    "src/ingestion/__main__.py",  # Entry point - tested manually
    "src/ingestion/ingest.py",  # CLI entry point - tested manually
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
